Moved auto populate function to own file
